refactored code: added header guards for .c.h modules
