+++ include/ast/nodes/identifier.h
@@ -1,7 +1,7 @@
 #include "ast.h"
 
-#ifndef AST_SYNTAX_IDENTIFIER_H
-#define AST_SYNTAX_IDENTIFIER_H
+#ifndef AST_NODES_IDENTIFIER_H
+#define AST_NODES_IDENTIFIER_H
 
 void AST_Identifier_free(AST_Identifier_t *identifier);
