+++ include/ast/nodes/statements.h
@@ -1,7 +1,7 @@
 #include "ast.h"
 
-#ifndef AST_SYNTAX_STATEMENTS_H
-#define AST_SYNTAX_STATEMENTS_H
+#ifndef AST_NODES_STATEMENTS_H
+#define AST_NODES_STATEMENTS_H
 
 void AST_Statements_free(AST_Statements_t *statements);
