+++ src/ast/util.c.h
@@ -1,0 +1,9 @@
+#ifndef AST_UTIL_C_H
+#define AST_UTIL_C_H
+
+#include "util/module_stack.c.h"
+#include "util/procedure_map.c.h"
+
+#else
+    #warning re-inclusion of module 'ast/util.c.h'
+#endif
