--- src/functions/calc.c.h
@@ -1,9 +1,0 @@
-#ifndef FUNC_CLAC_C_H
-#define FUNC_CLAC_C_H
-
-#include "parser/parse.c.h"
-#include "parser/vartable.c.h"
-
-VarEntry func_calc(ParserArgList argslist);
-
-#endif
